// examples/gen/api/gateway/v1/session_pb.d.ts
export declare type LoginReply = Message<"gateway.v1.LoginReply"> & {
  /**
   * @generated from field: string player_id = 1;
   */
  playerId: string;

  /**
   * @generated from field: string token = 2;
   */
  token: string;

  /**
   * @generated from field: common.v1.PlayerSummary player = 3;
   */
  player?: PlayerSummary | undefined;
};

/**
 * Describes the message gateway.v1.LoginReply.
 * Use `create(LoginReplySchema)` to create a new message.
 */
export declare const LoginReplySchema: GenMessage<LoginReply>;
// ...
export enum KickedReason {
  /**
   * @generated from enum value: KICKED_REASON_UNSPECIFIED = 0;
   */
  UNSPECIFIED = 0,

  /**
   * @generated from enum value: KICKED_REASON_LOGGED_IN_ELSEWHERE = 1;
   */
  LOGGED_IN_ELSEWHERE = 1,

  /**
   * @generated from enum value: KICKED_REASON_SESSION_EXPIRED = 2;
   */
  SESSION_EXPIRED = 2,
}